requirements.txt
paddlepaddle
paddlenlp